+++ src/oihana/models/traits/AlterDocumentTrait.php
@@ -8,7 +8,7 @@ use DI\NotFoundException;
 use Psr\Container\ContainerExceptionInterface;
 use Psr\Container\NotFoundExceptionInterface;
 
-use oihana\enums\Alter;
+use oihana\models\enums\Alter;
 use oihana\models\enums\ModelParam;
 use oihana\models\traits\alters\AlterArrayCleanPropertyTrait;
 use oihana\models\traits\alters\AlterArrayPropertyTrait;